Split rss and rdf, maybe this is a bad idea
